tests/requirements-test.txt
pytest>=8.0
httpx>=0.27
pytest-asyncio>=0.23
python-jose[cryptography]>=3.3
anyio>=4.0
